sqlite_to_PAW/Parser/website/models.py
from . import db
from flask_login import UserMixin
from sqlalchemy.sql import func


class Note(db.Model):
    id = db.Column(db.Numeric, primary_key=True) # id товара
    title = db.Column(db.String(255))
    specific = db.Column(db.Text)
    location = db.Column(db.String(255))
    price = db.Column(db.Numeric)
    date = db.Column(db.DateTime(timezone=True), default=func.now())
    user_id = db.Column(db.Integer, db.ForeignKey('user.id'))


class User(db.Model, UserMixin):
    id = db.Column(db.Integer, primary_key=True)
    email = db.Column(db.String(150), unique=True)
    password = db.Column(db.String(150))
    first_name = db.Column(db.String(150))
    notes = db.relationship('Note')
