sqlite_to_PAW/Parser/requirements.txt
flask
Flask-SQLAlchemy
flask-login
bs4